tests/wpt/css-tests/css-color-3_dev/xhtml1/caret-color-016.xht

<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
<html xmlns="http://www.w3.org/1999/xhtml"><head><meta charset="utf-8" />
<title>CSS Basic User Interface Test: caret-color visited link computed value</title>
<link href="mailto:rego@igalia.com" rel="author" title="Manuel Rego Casasnovas" />
<link href="http://www.w3.org/TR/css3-ui/#caret-color" rel="help" />
<link href="https://www.w3.org/TR/css3-color/#color0" rel="help" />
<link href="https://www.w3.org/TR/selectors4/#link" rel="help" />
<meta content="may interact" name="flags" />
<meta content="Test checks that computed style of caret-color on visited links doesn't leak privacy information." name="assert" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
  a {
    font-size: 3em;
    font-weight: bold;
    width: 10em;
    padding: 10px;
    background: black; /* the color of a thin object like the caret is easier to see on a black background. */

    color: white;
    caret-color: lime;
  }

  a:link {
    color: yellow;
  }

  a:visited {
    caret-color: cyan;
  }
</style>
</head><body>
  <p>Before running this test, this <a href="./">link</a> must have been visited. It will have yellow text if this is not the case. If it its text is yellow, you need to navigate to this link first.
  </p><p><a href="./" contenteditable="" id="link">link</a></p>
  <div id="log"></div>

  <script>
    setup(
      function(){
        /* Helper to get the link into the browsing history.
           Using a relative path because some browsers only allow replaceState within the same domain. */
        current_url = window.location.href;
        history.replaceState({},"","./");
        history.replaceState({},"",current_url);
      });

    test(
      function(){
        var link = document.getElementById("link");
        assert_equals(window.getComputedStyle(link)["caret-color"], "rgb(0, 255, 0)");
    }, "The computed value of a visited link should be the same than a non-visited link");
  </script>

</body></html>
